(bug 33414) Fix SQL error on merge/delete with blocked user.
Field ipb_id is the block id, ipb_user is the user id.

It is likely that there are more users than blocks, and so it was
very common that this query failed and aborted the delete/merge
when a blocked user is involved as the query tried to tamper with
the primary key column of the blocking table.

Change-Id: I166f0042b2445a06a63c6e3f70a28feae3603f13