Remove references to user fields removed in MW 1.34
The archive, revision, filearchive, image, oldimage, recentchanges,
logging and ipblocks tables no longer have user fields, so
$fieldInfo[1] should not be set.

I retained the actorStage concept for the benefit of extensions that
handle the UserMergeAccountFields hook. For the core tables, the actor
stage is always SCHEMA_COMPAT_NEW, so the actor fields are always used.

This should have no effect unless an extension is doing something very
weird in that hook. The point is just to remove unused field metadata.

Bug: T278917
Change-Id: I5e3867317cf6e90bf47704f38b4418b66314971e